feat: favour user and track rounding error

+++ test/utils/Invariants.sol
@@ -82,9 +82,8 @@ library Invariants {
         // );
 
         return
-            IERC20(mToken_).balanceOf(wrappedMToken_) >=
-            IWrappedMToken(wrappedMToken_).totalSupply() +
-                IWrappedMToken(wrappedMToken_).totalAccruedYield() +
+            int256(IERC20(mToken_).balanceOf(wrappedMToken_)) >=
+            int256(IWrappedMToken(wrappedMToken_).totalSupply() + IWrappedMToken(wrappedMToken_).totalAccruedYield()) +
                 IWrappedMToken(wrappedMToken_).excess();
     }
